Refactor the user repository
