+++ src/repository/UserRepository.ts
@@ -1,18 +1,7 @@
-import {
-  Activity,
-  Bidding,
-  Dispute,
-  ExperienceInfo,
-  Gender,
-  Portfolio,
-  Project,
-  Review,
-  Service,
-  User,
-} from "@prisma/client";
+import { Gender, User } from "@prisma/client";
 import Repository from "./Repository";
 
-type CreatableUser = {
+export type CreatableUser = {
   firstName: string;
   lastName: string;
   avatarImageURL: string;
@@ -26,28 +15,28 @@ type CreatableUser = {
   };
 };
 
-export type Client = Omit<User, "loginId"> & {
-  reviews: Review[];
-  activities: Activity[];
-  createdProjects: Project[];
-  disputes: Dispute[];
-};
-
-export type Provider = Omit<User, "loginId"> & {
-  biddings: Bidding[];
-  experiences: ExperienceInfo[];
-  portfolios: Portfolio[];
-  activities: Activity[];
-  createdServices: Service[];
-  disputes: Dispute[];
-};
-
 export default class UserRepository extends Repository {
   constructor() {
     super();
   }
 
-  async updateProviderAvatarImageURL(url: string, id: string) {
+  protected async createUser(data: CreatableUser, role: "Client" | "Provider") {
+    const newUser = await this.prisma.user.create({
+      data: {
+        ...data,
+        auth: {
+          create: {
+            ...data.auth,
+            isActive: true,
+            role,
+          },
+        },
+      },
+    });
+    return newUser;
+  }
+
+  async updateUserAvatarImageURL(url: string, id: string) {
     const updatedProvider = await this.prisma.user.update({
       where: { id },
       data: {
@@ -55,38 +44,6 @@ export default class UserRepository extends Repository {
       },
     });
     return updatedProvider;
-  }
-
-  async createClient(data: CreatableUser) {
-    const newClient = await this.prisma.user.create({
-      data: {
-        ...data,
-        auth: {
-          create: {
-            ...data.auth,
-            isActive: true,
-            role: "Client",
-          },
-        },
-      },
-    });
-    return newClient;
-  }
-
-  async createProvier(data: CreatableUser) {
-    const newProvider = await this.prisma.user.create({
-      data: {
-        ...data,
-        auth: {
-          create: {
-            ...data.auth,
-            isActive: true,
-            role: "Provider",
-          },
-        },
-      },
-    });
-    return newProvider;
   }
 
   async getByEmail(email: string): Promise<User | null> {
@@ -97,59 +54,7 @@ export default class UserRepository extends Repository {
       },
     });
     if (!userAuth) return null;
-    if (userAuth.role === "Administrator") return null;
     return userAuth.User;
-  }
-
-  async getProviderById(id: string): Promise<Provider | null> {
-    const provider = await this.prisma.user.findUnique({
-      where: { id },
-      include: {
-        biddings: true,
-        experienceInfo: true,
-        portfolio: true,
-        providerActivities: true,
-        services: true,
-        disputes: true,
-      },
-    });
-    if (!provider) return null;
-    return {
-      ...provider,
-      activities: provider.providerActivities,
-      experiences: provider.experienceInfo,
-      portfolios: provider.portfolio,
-      createdServices: provider.services,
-    };
-  }
-
-  async getClientById(id: string): Promise<Client | null> {
-    const client = await this.prisma.user.findUnique({
-      where: { id },
-      include: {
-        clientActivities: true,
-        disputes: true,
-        reviews: true,
-        projects: {
-          include: {
-            category: {
-              select: {
-                name: true,
-                slug: true,
-              },
-            },
-          },
-        },
-      },
-    });
-    if (!client) return null;
-    return {
-      ...client,
-      activities: client.clientActivities,
-      reviews: client.reviews,
-      createdProjects: client.projects,
-      disputes: client.disputes,
-    };
   }
 
   async getByPhoneNumber(phoneNumber: string) {
