--- src/repository/index.ts
@@ -1,2 +1,0 @@
-export { default as AdministratorRepository } from "./AdministratorRepository";
-export { default as CategoryRepository } from "./CategoryRepository";
